fix: move from entities to ficha
